Bin script to copy .env.example to .env
Adds a script from the Reaction project template that will copy values
from the .env.example into a local .env file. Iterates through the
example and copies any values that do not exist in the local file so it
is OK to run against an existing .env and get the new entries.

Sets proper security on the local .env file ownership.

NOTE: Windows compatibility unknown.